server/requirements.txt
blinker==1.6.2
cachelib==0.10.2
cffi==1.16.0
click==8.1.7
cryptography==41.0.7
dnspython==2.4.2
Flask==2.3.3
Flask-PyMongo==2.3.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
packaging==23.2
pycparser==2.21
pycryptodome==3.19.0
pymongo==4.5.0
python-docx==1.1.0
typing_extensions==4.9.0
Werkzeug==2.3.7
